`Contract` should be in `core` module Refactor
