+++ emerald-core/src/contract/error.rs
@@ -11,30 +11,6 @@ pub enum Error {
     InvalidContract(String),
 }
 
-impl From<ethabi::Error> for Error {
-    fn from(err: ethabi::Error) -> Self {
-        Error::InvalidABI(format!("Invalid ABI {:?}", err))
-    }
-}
-
-impl From<ethabi::spec::Error> for Error {
-    fn from(err: ethabi::spec::Error) -> Self {
-        Error::InvalidABI(format!("Invalid ABI Spec {:?}", err))
-    }
-}
-
-impl From<ethabi::token::Error> for Error {
-    fn from(err: ethabi::token::Error) -> Self {
-        Error::InvalidABIToken(err)
-    }
-}
-
-impl From<ethabi::spec::param_type::Error> for Error {
-    fn from(err: ethabi::spec::param_type::Error) -> Self {
-        Error::InvalidABI(format!("Invalid ABI Param {:?}", err))
-    }
-}
-
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match *self {
