+++ emerald-core/src/core/error.rs
@@ -15,7 +15,7 @@ pub enum Error {
 
     /// An invalid length
     InvalidLength(usize),
-std::convert::From<ethabi::spec::Error>
+
     /// An unexpected hexadecimal prefix (should be '0x')
     InvalidHexLength(String),
 
@@ -24,6 +24,30 @@ std::convert::From<ethabi::spec::Error>
 
     /// ECDSA crypto error
     EcdsaCrypto(secp256k1::Error),
+}
+
+impl From<ethabi::Error> for Error {
+    fn from(err: ethabi::Error) -> Self {
+        Error::InvalidABI(format!("Invalid ABI {:?}", err))
+    }
+}
+
+impl From<ethabi::spec::Error> for Error {
+    fn from(err: ethabi::spec::Error) -> Self {
+        Error::InvalidABI(format!("Invalid ABI Spec {:?}", err))
+    }
+}
+
+impl From<ethabi::token::Error> for Error {
+    fn from(err: ethabi::token::Error) -> Self {
+        Error::InvalidABIToken(err)
+    }
+}
+
+impl From<ethabi::spec::param_type::Error> for Error {
+    fn from(err: ethabi::spec::param_type::Error) -> Self {
+        Error::InvalidABI(format!("Invalid ABI Param {:?}", err))
+    }
 }
 
 impl From<hex::FromHexError> for Error {
